requirements.txt
Flask==1.0.2
pytest==3.6.3
python-slugify==1.2.6
redis==2.10.6
requests==0.9.0
WTForms==2.2.1
